refactor(project): separate integration from services
