+++ src/components/EditCardPaymentForm/EditCardPaymentForm.tsx
@@ -16,7 +16,7 @@ import styles from './EditCardPaymentForm.module.scss';
 import useForm from '#src/hooks/useForm';
 import { useAccountStore } from '#src/stores/AccountStore';
 import AccountController from '#src/stores/AccountController';
-import { getModule } from '#src/modules/container';
+import { getModule } from '#src/container';
 
 type Props = {
   onCancel: () => void;
