+++ src/containers/AccountModal/forms/ResetPassword.tsx
@@ -14,7 +14,7 @@ import LoadingOverlay from '#components/LoadingOverlay/LoadingOverlay';
 import { addQueryParam, removeQueryParam } from '#src/utils/location';
 import { logDev } from '#src/utils/common';
 import AccountController from '#src/stores/AccountController';
-import { getModule } from '#src/modules/container';
+import { getModule } from '#src/container';
 
 type Prop = {
   type: 'confirmation' | 'forgot' | 'reset' | 'edit';
