+++ src/containers/Layout/Layout.tsx
@@ -23,7 +23,7 @@ import { ACCESS_MODEL } from '#src/config';
 import { useProfileStore } from '#src/stores/ProfileStore';
 import { useProfiles, useSelectProfile } from '#src/hooks/useProfiles';
 import { IS_DEVELOPMENT_BUILD } from '#src/utils/common';
-import { getModule } from '#src/modules/container';
+import { getModule } from '#src/container';
 import ProfileController from '#src/stores/ProfileController';
 
 const Layout = () => {
