+++ src/containers/PlayerContainer/PlayerContainer.tsx
@@ -8,7 +8,7 @@ import type { JWPlayer } from '#types/jwplayer';
 import useContentProtection from '#src/hooks/useContentProtection';
 import LoadingOverlay from '#components/LoadingOverlay/LoadingOverlay';
 import ApiController from '#src/stores/ApiController';
-import { getModule } from '#src/modules/container';
+import { getModule } from '#src/container';
 
 type Props = {
   item: PlaylistItem;
