+++ src/hooks/series/useEpisodes.ts
@@ -4,7 +4,7 @@ import type { EpisodesWithPagination } from '#types/series';
 import type { Pagination } from '#types/pagination';
 import { CACHE_TIME, STALE_TIME } from '#src/config';
 import ApiController from '#src/stores/ApiController';
-import { getModule } from '#src/modules/container';
+import { getModule } from '#src/container';
 
 const getNextPageParam = (pagination: Pagination) => {
   const { page, page_limit, total } = pagination;
