+++ src/hooks/series/useSeries.ts
@@ -4,7 +4,7 @@ import type { Series } from '#types/series';
 import type { ApiError } from '#src/utils/api';
 import { CACHE_TIME, STALE_TIME } from '#src/config';
 import ApiController from '#src/stores/ApiController';
-import { getModule } from '#src/modules/container';
+import { getModule } from '#src/container';
 
 export const useSeries = (
   seriesId: string | undefined,
