+++ src/hooks/series/useSeriesLookup.ts
@@ -2,7 +2,7 @@ import { useQuery } from 'react-query';
 
 import { STALE_TIME, CACHE_TIME } from '#src/config';
 import ApiController from '#src/stores/ApiController';
-import { getModule } from '#src/modules/container';
+import { getModule } from '#src/container';
 
 export const useSeriesLookup = (mediaId: string | undefined) => {
   const apiController = getModule(ApiController);
