+++ src/hooks/useEntitlement.ts
@@ -8,7 +8,7 @@ import { isLocked } from '#src/utils/entitlements';
 import { useConfigStore } from '#src/stores/ConfigStore';
 import { useAccountStore } from '#src/stores/AccountStore';
 import CheckoutController from '#src/stores/CheckoutController';
-import { getModule } from '#src/modules/container';
+import { getModule } from '#src/container';
 
 export type UseEntitlementResult = {
   isEntitled: boolean;
