+++ src/hooks/useNotifications.ts
@@ -5,7 +5,7 @@ import { simultaneousLoginWarningKey } from '#components/LoginForm/LoginForm';
 import { queryClient } from '#src/containers/QueryProvider/QueryProvider';
 import { addQueryParams, removeQueryParamFromUrl } from '#src/utils/formatting';
 import AccountController from '#src/stores/AccountController';
-import { getModule } from '#src/modules/container';
+import { getModule } from '#src/container';
 import { useAccountStore } from '#src/stores/AccountStore';
 
 enum NotificationsTypes {
