+++ src/hooks/usePlaylist.ts
@@ -7,7 +7,7 @@ import { isScheduledOrLiveMedia } from '#src/utils/liveEvent';
 import { isTruthyCustomParamValue } from '#src/utils/common';
 import type { ApiError } from '#src/utils/api';
 import ApiController from '#src/stores/ApiController';
-import { getModule } from '#src/modules/container';
+import { getModule } from '#src/container';
 
 const placeholderData = generatePlaylistPlaceholder(30);
