+++ src/index.tsx
@@ -4,7 +4,7 @@ import 'wicg-inert';
 import { registerSW } from 'virtual:pwa-register';
 import 'reflect-metadata'; // include once in the app for inversify (see: https://github.com/inversify/InversifyJS/blob/master/README.md#-installation)
 
-import '#src/modules/init2';
+import './container';
 import App from './App';
 
 const rootElement = document.getElementById('root');
